requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
audiosegment==0.23.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.6
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==4.1.1
docker-pycreds==0.4.0
einops==0.2.0
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.101.1
fastdup==1.39
ffmpy==0.3.1
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gast==0.4.0
gdown==4.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==2.4.0
gradio==3.39.0
gradio_client==0.4.0
grpcio==1.57.0
grpcio-tools==1.48.2
h11==0.14.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.16.4
hyperframe==5.2.0
idna==2.10
imageio==2.31.3
imgaug==0.4.0
ipykernel==6.22.0
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
libretranslatepy==2.1.1
librosa==0.9.2
linkify-it-py==2.0.2
llvmlite==0.40.1
lmdb==1.4.1
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.1.1
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.1
numpy==1.23.0
oauthlib==3.2.2
open-clip-torch==2.20.0
opencv-python==4.8.0.76
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
orjson==3.9.5
outcome==1.2.0
packaging==23.0
pandas==1.5.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pillow-heif==0.13.0
platformdirs==3.5.0
pooch==1.6.0
portalocker==2.8.2
prefetch-generator==1.0.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.0.2
qdrant-client==1.4.0
rapidfuzz==3.3.0
referencing==0.30.2
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.9.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.5
sqlparse==0.4.2
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.8.30
timm==0.9.7
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
translate==3.6.1
trio==0.22.0
trio-websocket==0.10.2
typing==3.7.4.3
typing_extensions==4.7.1
uc-micro-py==1.0.2
urllib3==1.26.14
uvicorn==0.23.2
uvloop==0.17.0
vietocr==0.3.11
wandb==0.14.0
watchfiles==0.20.0
wcwidth==0.2.6
webrtcvad==2.0.10
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
